requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
Werkzeug==1.0.1
wikipedia==1.4.0
